requirements.txt
Scrapy==1.6.0
scrapy-mongodb==0.12.0
bandcamp-downloader